Replace the "ImageAnalysis" with "TileAnalysor"
Still with bugs
